+++ cuda/src/kernels/histogram_kernels.cu
@@ -9,18 +9,18 @@
 
 using namespace std;
 
-__global__ void init_DpE_kernel(double *DpE, int N_PART, int BINS) {
+__global__ void init_DpE_kernel(float *DpE, int N_PART, int BINS) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (i >= 2 * BINS) return;
 
-    double numerator = 6.0E-26 * N_PART;
-    double denominator = 5.24684E-24 * sqrt(2.0 * M_PI);
-    double exponent = -pow(3.0e-23 * (1.0 * i / BINS - 0.999) / 5.24684E-24, 2) / 2;
+    float numerator = 6.0E-26 * N_PART;
+    float denominator = 5.24684E-24 * sqrt(2.0 * M_PI);
+    float exponent = -pow(3.0e-23 * (1.0 * i / BINS - 0.999) / 5.24684E-24, 2) / 2;
     DpE[i] = (numerator / denominator) * exp(exponent);
 }
 
-__global__ void init_DxE_kernel(double *DxE, int N_PART, int BINS) {
+__global__ void init_DxE_kernel(float *DxE, int N_PART, int BINS) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= (BINS + 1) << 1) return;
     if (i < 2) {
@@ -53,12 +53,12 @@ __device__ uint32_t xorshift32(uint32_t* seed) {
     return x;
 }
 
-__device__ double d_xorshift(uint32_t *seed) {
+__device__ float d_xorshift(uint32_t *seed) {
     uint32_t x = xorshift32(seed);
-    return (double)x / (double)UINT32_MAX;
+    return (float)x / (float)UINT32_MAX;
 }
 
-__global__ void init_x_kernel(double* x, uint32_t base_seed, int N_PART) {
+__global__ void init_x_kernel(float* x, uint32_t base_seed, int N_PART) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     if (idx >= N_PART) return;
 
@@ -66,19 +66,19 @@ __global__ void init_x_kernel(double* x, uint32_t base_seed, int N_PART) {
     x[idx] = d_xorshift(&seed) * 0.5;
 }
 
-__global__ void init_p_kernel(double* p,  uint32_t base_seed, int N_PART) {
+__global__ void init_p_kernel(float* p,  uint32_t base_seed, int N_PART) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     if (idx >= (N_PART >> 1)) return;
     
     uint32_t seed = generate_random(base_seed);
 
     // Generate two random values using XORShift
-    double randomValue1 = (double)(xorshift32(&seed)) / UINT32_MAX;
-    double randomValue2 = (double)(xorshift32(&seed)) / UINT32_MAX;
+    float randomValue1 = (float)(xorshift32(&seed)) / UINT32_MAX;
+    float randomValue2 = (float)(xorshift32(&seed)) / UINT32_MAX;
 
     // Box-Muller transform to generate two normally distributed random numbers
-    double xi1 = sqrt(-2.0 * log(randomValue1 + 1E-35));
-    double xi2 = 2.0 * M_PI * randomValue2;
+    float xi1 = sqrt(-2.0 * log(randomValue1 + 1E-35));
+    float xi2 = 2.0 * M_PI * randomValue2;
 
     // Store the generated values in the p array
     p[2 * idx] = xi1 * cos(xi2) * 5.24684E-24;
@@ -86,7 +86,7 @@ __global__ void init_p_kernel(double* p,  uint32_t base_seed, int N_PART) {
     
 }
 
-__global__ void update_histograms_kernel(double *x, double *p, int *h, int *g, int *hg, int N_PART, int BINS) {
+__global__ void update_histograms_kernel(float *x, float *p, int *h, int *g, int *hg, int N_PART, int BINS) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
     if (idx >= N_PART) return;
 
@@ -104,15 +104,15 @@ __global__ void update_histograms_kernel(double *x, double *p, int *h, int *g, i
 
 
 // Kernel function to update positions and momenta
-__global__ void simulate_particle_motion(int number_of_steps, double *x, double *p, double *DxE, double *DpE, int *h, int *g, int *hg, int N_PART, double DT, double M, double sigmaL, double alfa, double pmin, double pmax) {
+__global__ void simulate_particle_motion(int number_of_steps, float *x, float *p, float *DxE, float *DpE, int *h, int *g, int *hg, int N_PART, float DT, float M, float sigmaL, float alfa, float pmin, float pmax) {
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (idx >= N_PART) return;
 
     uint32_t seed = idx + blockIdx.x + threadIdx.x * 31;
 
-    double x_tmp = x[idx];
-    double p_tmp = p[idx];
+    float x_tmp = x[idx];
+    float p_tmp = p[idx];
 
     int signop, k;
 
@@ -124,11 +124,11 @@ __global__ void simulate_particle_motion(int number_of_steps, double *x, double 
 
         if (k == 0) continue;
 
-        double randomValue = d_xorshift(&seed);
-        double xi1 = sqrt(-2.0 * log(randomValue + 1E-35));
+        float randomValue = d_xorshift(&seed);
+        float xi1 = sqrt(-2.0 * log(randomValue + 1E-35));
         randomValue = d_xorshift(&seed);
-        double xi2 = 2.0 * M_PI * randomValue;
-        double deltaX = sqrt(fabsf(k)) * xi1 * cos(xi2) * sigmaL;
+        float xi2 = 2.0 * M_PI * randomValue;
+        float deltaX = sqrt(fabsf(k)) * xi1 * cos(xi2) * sigmaL;
 
         deltaX = (fabs(deltaX) > 1.0 ? 1.0 * copysign(1.0, deltaX) : deltaX);
         x_tmp = (k % 2 ? -1.0 : 1.0) * (x_tmp - k) + deltaX;
@@ -142,9 +142,9 @@ __global__ void simulate_particle_motion(int number_of_steps, double *x, double 
         // TODO: Ask
 
         // for (int l = 1; l <= labs(k); ++l) {
-        double DeltaE = alfa * (p_tmp - pmin) * (pmax - p_tmp);
+        float DeltaE = alfa * (p_tmp - pmin) * (pmax - p_tmp);
         randomValue = d_xorshift(&seed);
-        double value = p_tmp * p_tmp + DeltaE * (randomValue - 0.5);
+        float value = p_tmp * p_tmp + DeltaE * (randomValue - 0.5);
         if (value < 0) {
             // Precision error. 
             // If p_tmp * p_tmp + DeltaE * (randomValue - 0.5) is negative, then the square root will be NaN.
@@ -167,8 +167,8 @@ __global__ void simulate_particle_motion(int number_of_steps, double *x, double 
 }
 
 // CUDA kernel for energy sum calculation
-__global__ void energy_sum_kernel(double *p, double *partialSum, int N_PART) {
-    extern __shared__ double sharedData[];
+__global__ void energy_sum_kernel(float *p, float *partialSum, int N_PART) {
+    extern __shared__ float sharedData[];
     int tid = threadIdx.x;
     int i = blockIdx.x * blockDim.x + threadIdx.x;
